Attack animation for rocket glove implemented
Rocket glove is now attached to the player camera and when the player shoots it plays the attack animation. However some refactoring must be made. The code needs to be cleaned and better planed. Also the animation needs some adjustments, like a second  part where the glove comes back and there should not be any delay when punching.