Biggining of the weapon related scripts refactor
The code for the weapon management and usage needs to be refactored and better planed. So in this commit a ScriptableObject for WeaponData and a WeaponInstance script have been implemented. he first one allows the creation of sccritableObjects for every weapon that will be in the game, wich hold the base information of them, and the second takes the scriptableObjects info and updates it according ti the player current progress in the game (progress = upgrades).