RocketGloves Prefab updated
Now the player has 2 RocketGloves, one in each hand, and punches alternately.
Each glove has its one animation. Each one has 4 animation events, at the begining it calls the EnableHitbox() and DisableShoot() functions, qhen the punch is complete it calls DisableHitBox() and at the end of the animation EnableShoot() is called.

EnableHitbox and DisableHitbox affect the hitboxes in the gloves.

EnableShoot and DisableShoot controll if the shoot function can execute or not according to the current state of the gloves. If one is in the midle of an animation then the other one can't start